finish demo; add test; add doc page
