+++ .gitattributes
@@ -1,1 +1,3 @@
 .git_archival.txt  export-subst
+
+*.mp4 binary
